cleaned up the code for models and routes

+++ routes/category.js
@@ -6,6 +6,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const router = express.Router();
 const Category = require('../models/category');
+const Department = require('../models/department');
 
 router.use(bodyParser.json());
 
@@ -16,6 +17,7 @@ router.post("/category/create", async (req, res) => {
     const existingCat = await Category.findOne({
       title: req.body.title
     });
+
     if (existingCat === null) {
       const newCat = new Category({
         title: req.body.title,
@@ -25,8 +27,7 @@ router.post("/category/create", async (req, res) => {
       await newCat.save();
       res.json({
         message: `Category added in ${req.body.departmentId}`,
-        title: req.body.title,
-        description: req.body.description
+        newCat
       });
     } else {
       res.status(400).json({
@@ -47,10 +48,11 @@ router.post("/category/create", async (req, res) => {
 router.get("/category", async (req, res) => {
   try {
     const categories = await Category.find().populate("department");
+
     res.json(categories);
   } catch (error) {
     res.status(400).json({
-      message: "An error occurred"
+      message: error.message
     });
   }
 });
@@ -61,25 +63,23 @@ router.get("/category", async (req, res) => {
 router.post("/category/update", async (req, res) => {
   try {
     const category = await Category.findById(req.query.id);
-    const departmentLink = await Department.findOne({
-      title: req.body.department
-    });
+    const department = await Department.findById(req.body.department);
     const oldCat = category.title;
-    if (category !== null) {
+
+    if (category && department) {
       category.title = req.body.title;
       category.description = req.body.description;
-      category.department = departmentLink;
+      category.department = req.body.department;
+
       await category.save();
       res.json({
         message: `Category ${oldCat} has been modified`,
-        title: req.body.title,
-        description: req.body.description,
-        department: req.body.department
+        category
       });
     } else {
       res.status(400).json({
         error: {
-          message: "Bad request"
+          message: "Category or department not found"
         }
       });
     }
@@ -95,19 +95,25 @@ router.post("/category/update", async (req, res) => {
 router.post("/category/delete", async (req, res) => {
   try {
     const category = await Category.findById(req.query.id);
-    await Category.deleteOne({
-      _id: req.query.id
-    });
-    await Product.deleteMany({
-      category: req.query.id
-    });
-    res.json({
-      message: `Deleted ${category.title} and all its products`
-    });
+    if (category) {
+      await Category.deleteOne({
+        _id: req.query.id
+      });
+      await Product.deleteMany({
+        category: req.query.id
+      });
+      res.json({
+        message: `Deleted ${category.title} and all its products`
+      });
+    } else {
+      res.status(400).json({
+        message: "Category not found"
+      });
+    }
   } catch (error) {
     res.status(400).json({
       error: {
-        message: "An error occurred"
+        message: error.message
       }
     });
   }
